Implement user edit, delete created events and refactor event code
Add edit, update and destroy actions to the event controller which
execute the desired actions if the user is the event creator and
redirects the user if not. Add the edit view with a form to update
the events and a link and a button in the show view to edit and
delete the events. Also refactor event related code fixing some
redirects/renders, adding an ordering in the index, renaming some
actions and hiding the subscribe/unsubscribe button for unsigned
users.